05_convolutional_neural_networks/assignments/transfer_learning/requirements.txt
# Transfer Learning -- Module 05
torch>=2.0.0
torchvision>=0.15.0
tensorboard>=2.12.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pyyaml>=6.0
Pillow>=10.0.0
